:recycle: [open-formulieren/open-forms#957] Refactor login buttons
