+++ src/components/FormStart/index.js
@@ -13,43 +13,25 @@ import { Toolbar, ToolbarList } from 'components/Toolbar';
 import useQuery from 'hooks/useQuery';
 import Types from 'types';
 import LoginButton, {START_FORM_QUERY_PARAM} from 'components/LoginButton';
-
-
-const LoginButtonIcons = ({ loginOptions }) => {
-  const optionsWithIcons = loginOptions.filter(option => option.logo && option.logo.imageSrc);
-  return (
-    <>
-      {optionsWithIcons.map(option => (
-        <Button
-          variant="image"
-          component="input"
-          type="image"
-          href={option.logo.href}
-          src={option.logo.imageSrc}
-          key={option.identifier}
-        />
-      ))}
-    </>
-  );
-};
-
-LoginButtonIcons.propTypes = {
-  loginOptions: PropTypes.arrayOf(PropTypes.shape({
-    identifier: PropTypes.string.isRequired,
-    label: PropTypes.string.isRequired,
-    url: PropTypes.string.isRequired,
-    logo: PropTypes.shape({
-      title: PropTypes.string.isRequired,
-      imageSrc: PropTypes.string.isRequired,
-      href: PropTypes.string,
-    }),
-  })).isRequired,
-};
+import LoginButtonIcons from 'components/LoginButtonIcons';
 
 
 const useStartSubmission = () => {
   const query = useQuery();
   return !!query.get(START_FORM_QUERY_PARAM);
+};
+
+
+const getLoginUrl = (loginOption) => {
+  const nextUrl = new URL(window.location.href);
+
+  const queryParams = Array.from(nextUrl.searchParams.keys());
+  queryParams.map(param => nextUrl.searchParams.delete(param));
+
+  nextUrl.searchParams.set(START_FORM_QUERY_PARAM, '1');
+  const loginUrl = new URL(loginOption.url);
+  loginUrl.searchParams.set('next', nextUrl.toString());
+  return loginUrl.toString();
 };
 
 
@@ -124,7 +106,13 @@ const FormStart = ({ form, onFormStart }) => {
               )
             }
             {
-              form.loginOptions.map((option) => <LoginButton option={option} key={option.identifier} />)
+              form.loginOptions.map((option) => (
+                <LoginButton
+                  key={option.identifier}
+                  option={option}
+                  getLoginUrl={getLoginUrl}
+                />
+              ))
             }
           </ToolbarList>
